Commit №3
Now, if the user runs the application without location tracking and then turns it on, location updates occur, and the weather data is loaded. The RecyclerView for the hourly forecast now displays only the forecast for the next 24 hours. The logic for handling daily forecast item clicks is not used but hasn't been removed. Forecast statuses are now string resources, and weather icons have also been added.