Commit №4
Support for a service to find cities and retrieve their data for further weather forecasts for these cities has been added. It is now possible to search for cities and add them to the database in order to load weather forecasts for them. The ability to get the weather forecast for the user's current location has been temporarily disabled. Certain functions have been moved from the main activity to fragments, and navigation has been added.