Commit №7
Now the application icons and background change depending on the time of day of the current city.
Fixed a bug with duplicating cities in the list in the cities manager fragment.
A feature to choose between temperature units has been added.
A feature to set weather forecast update frequency has been added."